Makes dompdf use get_url_contents() if it exists (for Pancake).

+++ src/FontMetrics.php
@@ -263,7 +263,12 @@ class FontMetrics
             }
             
             // Save the changes
-            file_put_contents($localFile, file_get_contents($remoteFile, null, $context));
+            $method = function_exists("get_url_contents") ? "get_url_contents" : "file_get_contents";
+            $remoteFileContent = $method($remoteFile, null, $context);
+            if (false === $remoteFileContent) {
+                return false;
+            }
+            file_put_contents($localFile, $remoteFileContent);
             
             if ( !file_exists($localFile) ) {
                 unlink("$cacheEntry.ufm");
